+++ CHANGELOG.md
@@ -16,6 +16,7 @@
     - `GridRotation::ALL_BUT_REFLECTIONS`, as `GridRotation::ALL` but excluding reflections.
     - `StandardCameras::world_space()`, so that clients don't need to consult the `Character` each frame.
     - `Universe::get_any()` allows looking up universe members without knowing their type.
+    - `UniverseTransaction::insert()` allows inserting objects into a `Universe` via transaction rather than directly.
     - Most `Listener` implementations now also implement `Clone` and `Debug`. This allows taking a listener and registering it with more than one `Notifier`. Relatedly, some uses of `-> impl Listener` have been replaced with concrete types.
 
 ### Changed